fix dropbox conflicts
